+++ R/generate_peels_network_internal.R
@@ -11,7 +11,7 @@
 #' @details  This function generates networks matching the 5 types described in Peel et al 2019(\url{https://doi.org/10.1073/pnas.1713019115}). All networks have 40 nodes,
 #'  160 edges, two node classes and four node sub-classes. The connections between the are equal across all 5 types.
 #'  As a result all networks generated have identical assortativity. However, as the sub-classes have different connection
-#'  probability the structures produced by the networks are very different. When projected into SETSe space the network types
+#'  probability the structures produced by the networks are very different. When projected into setse space the network types
 #'  occupy there own area, see Bourne 2020 for details
 #'  
 #' @return An igraph object that matches one of the 5 Peel's quintet types. The nodes are labelled with class and sub class.