+++ NEWS.md
@@ -1,3 +1,10 @@
+# rsetse 0.5
+
+* Added high dimensional functionality in the form of the setse_auto_hd function
+* Modified and re-named the prepare_ functions to make them appropriate for high dimensional embeddings
+* Created a new preparation function for the edges and removed that aspect from the original functions, making the functions simpler and more intuitive
+* Changed all "setse_" embedding functions from "SETSe_" to lowercase.
+
 # rSETSe 0.4
 
 * Added a `NEWS.md` file to track changes to the package.